bugfix: 違うファイルに、違う実装をしていたので、内容を交換。
